cmd: Add the root command
