+++ types/imageinspection.go
@@ -6,6 +6,13 @@ import (
 	"github.com/opencontainers/go-digest"
 )
 
+const (
+	// PivotDonePath is the path to the file used to denote pivot work
+	PivotDonePath = "/etc/os-container-pivot.stamp"
+	// PivotName is literally the name of the new pivot
+	PivotName = "ostree-container-pivot"
+)
+
 // ImageInspection is a public implementation of
 // https://github.com/projectatomic/skopeo/blob/master/cmd/skopeo/inspect.go#L20-L31
 type ImageInspection struct {
